requirements.txt
flask==2.0.2
Werkzeug==2.0.3
jinja2==3.0.2
flask_login==0.5.0
flask_migrate==3.1.0
WTForms==3.0.0
flask_wtf==1.0.0
flask_sqlalchemy==2.5.1
sqlalchemy==1.4.29
email_validator==1.1.3
gunicorn==20.1.0
flask-restx==0.5.1
python-dotenv==0.19.2
Flask-Minify==0.37
requests
Pillow
numpy
matplotlib
plotly
scipy
tqdm
imageio
ftfy
torch
torchvision
imgui
glfw
pyopengl
imageio-ffmpeg
pyspng
albumentations
nibabel
numpy
pydicom
matplotlib
scipy
scikit-image
scikit-learn
plotly
h5py
nibabel
SimpleITK
opencv-python
tensorflow
nltk
seaborn
